Reprojection of MODIS datasets
The MODIS reprojection now works. It returns a fairly hairy dictionary
indexed by product family (MCD43A1 and A2), which is then indexed by
S2 dates, and which contains the relevant files that can be read in
using GDAL. The files have the same projection and spatial extent
as the S2 tile, but far fewer pixels (obviously!).