Mostly working version for MCD43 comparisons
This version works on the different products, and is able to provide
the user with the S2 BOA reflectance convolved with a Gaussian PSF,
provided on the same grid as the relevant MODIS scene. The MCD43
product is used to provide a prediction of the S2 convolved BOA
reflectance. There are currently a number of issues:

* I'm afraid you need to fix how the modis bands are used to predict S2
bands. Basically, for each S2 band, you can either find the equivalent
MODIS band (with a weight of 1), or provide a set of bands with their
own weights. This is done in `extract_data.py` L397
* I think I have only selected high quality MCD43A1 pixels, but I'm not
sure
* The BC comparison has an issue as the files have the wrong geocoding. I
can fix that quickly, but ignore it for the time being (we should just
be able to run the same analyses on the BC data once I fix that)
* I haven't implemented individual S2 product masking
* I have convolved with PSF, but haven't optimised for shift
Angles of S2 scene are assumed constant over scene